use 64-bit integers by in MLton
MLton seems to default to 32-bit integers, which is really quite small,
small enough that I was frequently seeing overflows on the example
programs I tested (Fibonacci, factorial, etc.). Of course, this change
merely raises the threshold before you see overflows, it doesn't make
them go away, but I don't consider this to be a correctness issue
because programs with the Haskell interpreter would have hung as soon as
they tried to operate on or display these numbers anyway. The fact that
we have a special optimization for arithmetic is incidental.