support compiling with SMLnj
The heap images produced by SMLnj have a different extension depending
on the architecture and operating system, so I'm not sure what the best
way to make them work properly with Make is. As it is, we use a phony
target that calls `ml-build` and leave it at that.

We also have to change main to return an exit code rather than `unit` so
it plays nicely with how SMLnj wants exported functions to work.