ruscarpone: follow spec w.r.t. instruction lookup
Previously we used the principle that instruction lookup can fail if an
interpreter does not associate an operation with a particular symbol.
However, most interpreters do associate every operation with a symbol.
The only exception is the null interpreter and anything derived from it.
What happens if we attempt to discover what operation 0 assigns to a
string, such as with `[...]0*`? The language description says only this:

> There is a special interpreter in Mascarpone called "null". It is an
error to try to interpret anything with this interpreter. Expect that
any program that tries to do this will come crashing to a halt, or will
spin off into space and never be heard from again, or something equally
impressive.

In light of this, I think a better interpretation is that instruction
lookup can never fail, but it may yield an operation which fails when
performed.