ruscarpone: track when interp derives from initial
We can avoid the incorrect assumption, made by `Interpreter::show`, that
any interpreter having `Nop` as its fallback must have been derived from
the initial interpreter, and so its representation should start with `v`
and then list those operations which were changed from their definitions
in the initial interpreter. This assumption was made because there isn't
any better way to get this information. However, it meant that, if there
is an interpreter whose fallback is set is set to `Nop` explicitly, such
as through the code `[]v*1`, then it would be pretty-printed incorrectly.