ruscarpone: fix interpreter capture semantics
The spec says:

> The parent interpreter of the current interpreter is generally the
interpreter that was used to execute the current operation (that is, the
operation currently being interpreted; it consists of a string of
symbols is interpreted by the current interpreter.)

However, it doesn't describe exactly how the parent of the current
interpreter reassigned to maintain this generality. In particular, when
running an operation which has an interpreter captured by `*`, the
correct semantics are to set this interpreter as current, with its
parent set to the old current interpreter. Then after execution, we set
the current interpreter to the parent of the current interpreter. (This
is what the Haskell interpreter does.)

What we were doing instead is to replace the current interpreter and its
entire parent chain with the captured interpreter and its parent chain,
and after execution, to restore the old one.