ruscarpone: add debug print builtin
The fun thing about the '@' builtin is that it pushes an arbitrary
interpreter, and that interpreter can have arbitrary additional
functionality assigned to other operations. This gives us a place where
we can introduce more "secret" intrinsics without polluting the initial
namespace. In particular, we add a '?' operator which dumps the stack
state to STDERR. By default, this operator is in a separate "IO
interpreter", and this interpreter is what is pushed when expanding IO
operations.

Any program that wants permanent access to the debug print intrinsic can
simply begin with:

    v'.>@   (expand output operator)
    /$/$/$  (remove string, leaving only interpreter)
    '?>     (extract debug operator)
    v/'?<^  (add this operator to the active interpreter)